fix: Do not limit boot entry title length
In max-ishere/grubshin-bootpact#11, it became evident that real boot
entries are rather descriptive and long. The original expectation was
that someone will either rename the entries to be shorter, or that
distros make them short by default.

Since that is not the case, the boot entry was updated. The left edge is
roughly in the same spot. The right edge, however, is right on the
screen edge.

In testing, this is the best compromise between aestetics and
practicality. Long entries will look a bit weird, but acceptable. Short
entries will look weird as well, but also acceptable.

At the end of the day, the user can always tweak the width locally.